==> R/Modelling/1a_runAme_interactions.R <==
################
# workspace
if(Sys.info()['user']=='janus829' | Sys.info()['user']=='s7m'){
	source('~/Research/conflictEvolution/R/setup.R') }
if(Sys.info()['user']=='cassydorff' | Sys.info()['user']=='cassydorff'){
	source('~/ProjectsGit/conflictEvolution/R/setup.R') }
if(Sys.info()['user']=='maxgallop'){
	source('~/Documents/conflictEvolution/R/setup.R') }
loadPkg('devtools') ; devtools::install_github('s7minhas/amen') ; library(amen)
################

################
# load data
load(paste0(pathData, 'nigeriaMatList_acled_v7.rda')) # loads yList object
load(paste0(pathData, 'exoVars.rda')) # load xNodeL, xDyadL

# focus on post 2000 data [few actors beforehand]
yrs = char(2000:2016)
yList = yList[yrs] ; xDyadL = xDyadL[yrs] ; xNodeL = xNodeL[yrs]
###############

###############
# create interactions
xNodeL = lapply(xNodeL, function(x){
	govActor = grepl('Forces of Nigeria', rownames(x), fixed=TRUE)*1
	riotsProtestsAgainst_gov = govActor*x[,'riotsProtestsAgainst']
	riotsAgainst_gov = govActor*x[,'riotsAgainst']
	protestsAgainst_gov = govActor*x[,'protestsAgainst']
	vioCivEvents_gov = govActor*x[,'vioCivEvents']
	x = cbind(x,
		govActor=govActor,
		riotsProtestsAgainst_gov=riotsProtestsAgainst_gov,
		riotsAgainst_gov=riotsAgainst_gov,
		protestsAgainst_gov=protestsAgainst_gov,
		vioCivEvents_gov=vioCivEvents_gov
		)
	return(x)
})
###############

################
# set up model specs
subListArray = function(lA, vars, dims=2){
	if(dims==2){ return( lapply(lA, function(x){ x[,vars, drop=FALSE] }) ) }
	if(dims==3){ return( lapply(lA, function(x){ x[,,vars,drop=FALSE] }) ) } }
designArrays = list(
	base=list(
		senCovar=subListArray(xNodeL, c('riotsProtestsAgainst', 'vioCivEvents', 'groupSpread', 'govActor'), 2),
		recCovar=subListArray(xNodeL, c('riotsProtestsAgainst', 'vioCivEvents', 'groupSpread', 'govActor'), 2),
		dyadCovar=subListArray(xDyadL, c('postBoko', 'elecYear', 'ngbrConfCount'), 3)
		),
	baseInt=list(
		senCovar=subListArray(xNodeL, c(
			'riotsProtestsAgainst', 'riotsProtestsAgainst_gov', 
			'vioCivEvents', 'vioCivEvents_gov',
			'groupSpread', 'govActor'), 2),
		recCovar=subListArray(xNodeL, c(
			'riotsProtestsAgainst', 'riotsProtestsAgainst_gov', 
			'vioCivEvents', 'vioCivEvents_gov',
			'groupSpread', 'govActor'), 2),
		dyadCovar=subListArray(xDyadL, c('postBoko', 'elecYear', 'ngbrConfCount'), 3)
		),
	rioProSep=list(
		senCovar=subListArray(xNodeL, c('riotsAgainst', 'protestsAgainst','vioCivEvents', 'groupSpread', 'govActor'), 2),
		recCovar=subListArray(xNodeL, c('riotsAgainst', 'protestsAgainst','vioCivEvents', 'groupSpread', 'govActor'), 2),
		dyadCovar=subListArray(xDyadL, c('postBoko', 'elecYear', 'ngbrConfCount'), 3)
		)
	)

# parallelize model run
loadPkg(c('parallel','foreach'))
cores=length(designArrays) ; cl=makeCluster(cores) ; registerDoParallel(cl)
ameFits = foreach(i = 1:length(designArrays), .packages=c('amen')) %dopar% {
	fit = ame_repL(
		Y=yList, Xdyad=designArrays[[i]]$dyadCovar,
		Xrow=designArrays[[i]]$senCovar, Xcol=designArrays[[i]]$recCovar,
		symmetric=FALSE, rvar=TRUE, cvar=TRUE, R=2, 
		model='bin', intercept=TRUE, seed=6886,
		burn=10000, nscan=50000, odens=25,
		plot=FALSE, gof=TRUE, periodicSave=FALSE )
	return(fit)
} ; stopCluster(cl)
names(ameFits) = names(designArrays)
################

################
# save
save(
	ameFits, designArrays,
	file=paste0(pathResults, 'ameResults_interactions.rda')
	)
################
